split Application class
